requirements.txt
django>=5.0
djangorestframework
psycopg2-binary
python-dotenv
PyJWT
drf-spectacular
pytest
pytest-django